+++ argilla-server/CHANGELOG.md
@@ -17,6 +17,9 @@ These are the section headers that we use:
 ## [Extralit] [0.4.0](https://github.com/extralit/extralit/compare/v0.3.0...v0.4.0)
 ### Added
 
+### Changed
+
+- Changed the POST `/documents` endpoint to return Document schema instead of only the document id.
 
 
 ## [Argilla] [2.5.0](https://github.com/argilla-io/argilla/compare/v2.4.1...v2.5.0)